+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/server/AkkaHttpBasedRouteProvider.scala
@@ -88,21 +88,14 @@ class AkkaHttpBasedRouteProvider(
       _                    = logger.info(s"Designer config loaded: \nfeatureTogglesConfig: $featureTogglesConfig")
       countsReporter <- createCountsReporter(featureTogglesConfig, environment, sttpBackend)
       deploymentServiceSupplier = new DelayedInitDeploymentServiceSupplier
-      processCategoryService    = new ConfigProcessCategoryService(resolvedConfig)
       typeToConfigAndReload <- prepareProcessingTypeData(
         config,
         deploymentServiceSupplier,
-        processCategoryService,
         processingTypeDataProviderFactory,
         sttpBackend
       )
       (typeToConfig, reload) = typeToConfigAndReload
     } yield {
-      val stateDefinitionService = new ProcessStateDefinitionService(
-        typeToConfig.mapCombined(_.statusNameToStateDefinitionsMapping),
-        processCategoryService
-      )
-
       val analyticsConfig = AnalyticsConfig(resolvedConfig)
 
       val modelData = typeToConfig.mapValues(_.modelData)
@@ -173,7 +166,13 @@ class AkkaHttpBasedRouteProvider(
         dmDispatcher,
         deploymentService
       )
-      val testExecutorService = new ScenarioTestExecutorServiceImpl(scenarioResolver, dmDispatcher)
+      val testExecutorService    = new ScenarioTestExecutorServiceImpl(scenarioResolver, dmDispatcher)
+      def processCategoryService = typeToConfig.combined.categoryService
+
+      val stateDefinitionService = new ProcessStateDefinitionService(
+        typeToConfig.mapCombined(combined => (combined.statusNameToStateDefinitionsMapping, combined.categoryService)),
+      )
+
       val processService = new DBProcessService(
         deploymentService,
         newProcessPreparer,
@@ -215,9 +214,8 @@ class AkkaHttpBasedRouteProvider(
 
       val componentService = DefaultComponentService(
         ComponentLinksConfigExtractor.extract(resolvedConfig),
-        typeToConfig.mapCombined(_.componentIdProvider),
+        typeToConfig.mapCombined(combined => (combined.componentIdProvider, combined.categoryService)),
         processService,
-        processCategoryService,
         additionalUIConfigProvider
       )
 
@@ -353,7 +351,7 @@ class AkkaHttpBasedRouteProvider(
       tapirRelatedRoutes: List[Route],
       apiResourcesWithAuthentication: List[RouteWithUser],
       apiResourcesWithoutAuthentication: List[Route],
-      processCategoryService: ProcessCategoryService,
+      processCategoryService: => ProcessCategoryService,
       developmentMode: Boolean
   )(implicit executionContext: ExecutionContext): Route = {
     // TODO: In the future will be nice to have possibility to pass authenticator.directive to resource and there us it at concrete path resource
@@ -425,7 +423,6 @@ class AkkaHttpBasedRouteProvider(
   private def prepareProcessingTypeData(
       designerConfig: ConfigWithUnresolvedVersion,
       deploymentServiceSupplier: Supplier[DeploymentService],
-      categoriesService: ProcessCategoryService,
       processingTypeDataProviderFactory: ProcessingTypeDataProviderFactory,
       sttpBackend: SttpBackend[Future, Any]
   )(implicit executionContext: ExecutionContext): Resource[
@@ -440,7 +437,7 @@ class AkkaHttpBasedRouteProvider(
       .make(
         acquire = IO(
           BasicProcessingTypeDataReload.wrapWithReloader(() =>
-            processingTypeDataProviderFactory.create(designerConfig, deploymentServiceSupplier, categoriesService)
+            processingTypeDataProviderFactory.create(designerConfig, deploymentServiceSupplier)
           )
         )
       )(