+++ docs/Changelog.md
@@ -4,6 +4,7 @@
 -------------------------
 * [#4711](https://github.com/TouK/nussknacker/pull/4711) [#4862](https://github.com/TouK/nussknacker/pull/4862) Added AdditionalUIConfigProviderFactory API that allows changing components' configs and scenario properties' UI configs without model reload
 * [#4860](https://github.com/TouK/nussknacker/pull/4860) Rename `additionalProperties` to `scenarioProperties`
+* [#4854](https://github.com/TouK/nussknacker/pull/4854) Categories configuration redesign
 
 1.12.0 (Not released yet)
 -------------------------